+++ src/ui/src/service/utils.js
@@ -11,6 +11,7 @@
  */
 
 import merge from 'lodash/merge'
+import http from '@/api'
 
 /**
  * 根据是否开启count生成新参数
@@ -24,3 +25,89 @@ export const enableCount = (params = {}, flag = false) => {
 }
 
 export const onePageParams = () => ({ start: 0, limit: 1 })
+
+// 使用enableCount的方式滚动拉取接口数据（先获取总数）
+export const rollReqUseCount = async (
+  url,
+  params = {},
+  options = {},
+  config = {},
+  method = 'post'
+) => {
+  const { start = 1, limit = 1000, countKey = 'count', listKey = 'info' } = options
+
+  let index = start
+  const size = limit
+  const results = []
+
+  // 先获取到总数
+  const { [countKey]: total = 0 } = await http[method](url, enableCount(params, true), config)
+
+  // 分页方法
+  const page = index => ({ ...(params.page || {}), start: (index - 1) * size, limit: size })
+
+  // 请求列表的req
+  const req = index => http[method](url, enableCount({
+    ...params,
+    page: page(index)
+  }, false), config)
+
+  // 列表一共要拉取多少次
+  const max = Math.ceil(total / size)
+
+  // 循环组装得到所有的req
+  const reqs = []
+  while (index <= max) {
+    reqs.push(req(index))
+    index += 1
+  }
+
+  const all = await Promise.all(reqs)
+  all.forEach(({ [listKey]: list = [] }) => {
+    results.push(...list)
+  })
+
+  return results
+}
+
+// 常规方法滚动拉取接口数据（先拉取第1页）
+export const rollReq = async (
+  url,
+  params = {},
+  options = {},
+  config = {},
+  method = 'post'
+) => {
+  const { start = 1, limit = 1000, countKey = 'count', listKey = 'info' } = options
+
+  let index = start
+  const size = limit
+  const results = []
+
+  const page = index => ({ ...(params.page || {}), start: (index - 1) * size, limit: size })
+
+  const req = index => http[method](url, {
+    ...params,
+    page: page(index)
+  }, config)
+
+  // 先拉起始页，通常是第1页，同时得到总数
+  const { [countKey]: total = 0, [listKey]: list = [] } = await req(index)
+  results.push(...list)
+
+  // 一共要拉取多少次
+  const max = Math.ceil(total / size)
+
+  const reqs = []
+  while (index < max) {
+    index += 1
+    reqs.push(req(index))
+  }
+
+  const rest = await Promise.all(reqs)
+  rest.forEach(({ [listKey]: list = [] }) => {
+    results.push(...list)
+  })
+
+  return results
+}
